src/main.py
"""
Main entry point for the DEM-based Internal Erosion Model.
"""

import logging
import yaml
from pathlib import Path
import numpy as np
from src.dem.solver import DEMSolver
from src.cfd.solver import CFDSolver
from src.erosion.model import ErosionModel

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('simulation.log'),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

def load_config(config_path: str) -> dict:
    """Load configuration from YAML file."""
    try:
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        return config
    except Exception as e:
        logger.error(f"Failed to load configuration: {e}")
        raise

def initialize_simulation(config: dict):
    """Initialize simulation components based on configuration."""
    logger.info("Initializing simulation components...")
    
    # Initialize solvers
    dem_solver = DEMSolver(config)
    cfd_solver = CFDSolver(config)
    erosion_model = ErosionModel(config)
    
    # Initialize particles
    num_particles = config['simulation']['initial_particles']
    dem_solver.initialize_particles(num_particles)
    
    # Initialize fluid fields
    cfd_solver.initialize_fields()
    
    return dem_solver, cfd_solver, erosion_model

def run_simulation(dem_solver: DEMSolver, 
                  cfd_solver: CFDSolver, 
                  erosion_model: ErosionModel,
                  config: dict):
    """Run the coupled DEM-CFD-erosion simulation."""
    time_steps = config['simulation']['time_steps']
    output_interval = config['simulation']['output_interval']
    
    for step in range(time_steps):
        logger.info(f"Simulation step {step + 1}/{time_steps}")
        
        # Update fluid state
        cfd_solver.step()
        
        # Compute fluid forces on particles
        particle_positions = dem_solver.get_particle_data()['positions']
        fluid_forces = cfd_solver.compute_fluid_forces(particle_positions)
        
        # Update particle states
        dem_solver.step()
        
        # Compute erosion and update particle masses
        # TODO: Implement erosion updates
        
        # Output results if needed
        if (step + 1) % output_interval == 0:
            logger.info(f"Outputting results at step {step + 1}")
            # TODO: Implement result output

def main():
    """Main entry point for the simulation."""
    try:
        logger.info("Starting DEM-based Internal Erosion Model...")
        
        # Load configuration
        config_path = Path("config.yaml")
        config = load_config(config_path)
        
        # Initialize simulation
        dem_solver, cfd_solver, erosion_model = initialize_simulation(config)
        
        # Run simulation
        run_simulation(dem_solver, cfd_solver, erosion_model, config)
        
        logger.info("Simulation completed successfully.")
        
    except Exception as e:
        logger.error(f"Simulation failed: {e}")
        raise

if __name__ == "__main__":
    main()
