requirements.txt
numpy>=1.21.0
scipy>=1.7.0
matplotlib>=3.4.0
pytest>=7.0.0
pytest-cov>=3.0.0
pyvista>=0.34.0
pyyaml>=6.0
